requirements.txt
requests>=2.26.0
python-Levenshtein>=0.12.2
fastapi>=0.68.0
uvicorn>=0.15.0 